pyproject.toml
[project]
name = "agents-workshop"
version = "0.1.0"
description = "Hands-on workshop teaching 7 agentic AI patterns using Python, Azure OpenAI, and the Microsoft Agent Framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Workshop Team" }
]
keywords = [
    "agents",
    "ai",
    "mcp",
    "azure-openai",
    "workshop",
    "agentic-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Microsoft Agent Framework
    "agent-framework>=0.1.0",
    "azure-ai-projects>=1.0.0b1",
    "azure-identity>=1.15.0",
    
    # Azure OpenAI
    "openai>=1.0.0",
    
    # MCP Protocol
    "mcp>=1.0.0",
    
    # AG-UI Protocol
    "ag-ui-core>=0.1.0",
    
    # Web Framework (AG-UI/A2A servers)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    
    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "azure-monitor-opentelemetry>=1.2.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    
    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.28.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    
    # Type Checking
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",
    
    # Linting & Formatting
    "ruff>=0.1.14",
    
    # Pre-commit
    "pre-commit>=3.6.0",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/agents-workshop"
Documentation = "https://github.com/your-org/agents-workshop#readme"
Repository = "https://github.com/your-org/agents-workshop"
Issues = "https://github.com/your-org/agents-workshop/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "contract: marks tests as contract tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "agent_framework.*",
    "ag_ui.*",
    "mcp.*",
    "azure.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]
"notebooks/*" = ["E402", "ERA001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
